ensure bars don't create damage that is higher than current value
If player receives a damage higher than their current hp, that damage's width must be limited only to the current hp

+++ 2D Metroidvania/Assets/Scripts/Utilities/Constants.cs
@@ -128,5 +128,5 @@ public class Constants {
   public static float maxMPDisplayableLimit = 500;
 
   // ensures that each HP unit (until reaching 500) displays in 2 UI units
-  public static float containerMultiplier = 2;
+  public static int containerMultiplier = 2;
 }
